requirements.txt
face_recognition
dlib
opencv-python
pillow
numpy